Add files via upload
Task management is the art of organizing and prioritizing tasks to achieve goals effectively. It involves the entire lifecycle of a task, from initial planning and creation to completion and monitoring.
Task management is the process of effectively and efficiently tracking, managing, and executing the life cycle of a task or many tasks within a project, from inception to execution. The purpose of task management is to improve the decision-making, communication, efficiency, and effectiveness of a task or project.
Task management tools can be anything from a pen-and-paper planner to sophisticated software applications. The best tool depends on your individual needs and preferences.
# html #css #reactjs #nodejs #typescript #mongodb #javascript